Correct fix for bin.pack format repitition.
This partially reverts [1]. The "x" format specifier is also present in Lua
5.3. The problem was actually that bin.pack arguments were skipped on each
occurrence of "x". An additional problem I have also found is that repeated
formats (e.g. "c2") did not correctly increase the position in the arguments
(variable `i` in the code). There are now tests for both problems.

[1] https://svn.nmap.org/nmap-exp/patrick/nse-lua53@35923

git-svn-id: https://svn.nmap.org/nmap-exp/patrick/nse-lua53@35930 e0a8ed71-7df4-0310-8962-fdc924857419